move article parser and getter into the right services
