+++ server/services/prismic-parsers.js
@@ -2,9 +2,13 @@
 import {List} from 'immutable';
 import type {Exhibition} from '../content-model/exhibition';
 import getBreakpoint from '../filters/get-breakpoint';
-import {getPromo, asText, asHtml, prismicImageToPicture, getContributors} from './prismic-content';
+import {
+  getPromo, asText, asHtml, prismicImageToPicture, getContributors, getPublishedDate,
+  getFeaturedMediaFromBody, convertContentToBodyParts, getSeries
+} from './prismic-content';
 import type {DateRange} from '../content-model/content-blocks';
 import type {EventFormat} from '../content-model/event';
+import type {Article} from '../model/article';
 
 export function parseExhibitionsDoc(doc): Exhibition {
   const featuredImage = prismicImageToPicture({image: doc.data.featuredImage});
@@ -61,4 +65,72 @@ export function parseEventDoc(doc): Event {
   }: Event);
 
   return e;
+}
+
+export function parseArticleDoc(doc): Article {
+  // TODO : construct this not from strings
+  const url = `/articles/${doc.id}`;
+
+  const publishDate = getPublishedDate(doc);
+  const featuredMedia = getFeaturedMediaFromBody(doc);
+
+  // TODO: Don't convert this into thumbnail
+  const promo = doc.data.promo.find(slice => slice.slice_type === 'editorialImage');
+  const thumbnail = promo && prismicImageToPicture(promo.primary);
+  const description = promo && asText(promo.primary.caption); // TODO: Do not use description
+  const contributors = getContributors(doc);
+  const series = getSeries(doc);
+
+  const bodyParts = convertContentToBodyParts(doc.data.body);
+
+  // TODO: The whole scheduled content has some work to be getting on with
+  const seriesWithCommissionedLength = series.find(series => series.commissionedLength);
+  const positionInSeries = seriesWithCommissionedLength && (seriesWithCommissionedLength.positionInSeries || 1);
+
+  const article: Article = {
+    contentType: 'article',
+    headline: asText(doc.data.title),
+    url: url,
+    datePublished: publishDate,
+    thumbnail: thumbnail,
+    author: contributors,
+    series: series,
+    bodyParts: bodyParts,
+    mainMedia: [featuredMedia],
+    description: description,
+    positionInSeries: positionInSeries
+  };
+
+  return article;
+}
+
+export function parseWebcomicDoc(doc): Article {
+  // TODO : construct this not from strings
+  const url = `/articles/${doc.id}`;
+
+  // TODO: potentially get rid of this
+  const publishDate = getPublishedDate(doc);
+  const mainMedia = [prismicImageToPicture({ image: doc.data.image })];
+
+  // TODO: Don't convert this into thumbnail
+  const promo = doc.data.promo.find(slice => slice.slice_type === 'editorialImage');
+  const thumbnail = promo && prismicImageToPicture(promo.primary);
+  const description = asText(promo.primary.caption); // TODO: Do not use description
+  const contributors = getContributors(doc);
+  const series = getSeries(doc);
+
+  const article: Article = {
+    contentType: 'comic',
+    headline: asText(doc.data.title),
+    url: url,
+    datePublished: publishDate,
+    thumbnail: thumbnail,
+    author: contributors,
+    series: series,
+    bodyParts: [],
+    mainMedia: mainMedia,
+    description: description
+  };
+
+  return article;
 }
